checks: use register as a decorator
This makes the code simpler and matches currently expected behavior.

This triggered several changes in underlying code to document
parameters, reorganize and type annotate.
