+++ weblate/utils/celery.py
@@ -4,11 +4,16 @@
 
 """Celery integration helper tools."""
 
+from __future__ import annotations
+
 import os
+import time
+from collections import defaultdict
 
 from celery import Celery
 from celery.signals import after_setup_logger, task_failure
 from django.conf import settings
+from django.core.cache import cache
 from django.core.checks import run_checks
 
 # set the default Django settings module for the 'celery' program.
@@ -107,3 +112,51 @@ def get_task_progress(task):
 
     # Not yet started
     return 0
+
+
+def is_celery_queue_long():
+    """
+    Checks whether celery queue is too long.
+
+    It does trigger if it is too long for at least one hour. This way peaks are
+    filtered out, and no warning need be issued for big operations (for example
+    site-wide autotranslation).
+    """
+    from weblate.trans.models import Translation
+
+    cache_key = "celery_queue_stats"
+    queues_data = cache.get(cache_key, {})
+
+    # Hours since epoch
+    current_hour = int(time.time() / 3600)
+    test_hour = current_hour - 1
+
+    # Fetch current stats
+    stats = get_queue_stats()
+
+    # Update counters
+    if current_hour not in queues_data:
+        # Delete stale items
+        for key in list(queues_data.keys()):
+            if key < test_hour:
+                del queues_data[key]
+        # Add current one
+        queues_data[current_hour] = stats
+
+        # Store to cache
+        cache.set(cache_key, queues_data, 7200)
+
+    # Do not fire if we do not have counts for two hours ago
+    if test_hour not in queues_data:
+        return False
+
+    # Check if any queue got bigger
+    base = queues_data[test_hour]
+    thresholds = defaultdict(lambda: 50)
+    # Set the limit to avoid trigger on auto-translating all components
+    # nightly.
+    thresholds["translate"] = max(1000, Translation.objects.count() / 30)
+    return any(
+        stat > thresholds[key] and base.get(key, 0) > thresholds[key]
+        for key, stat in stats.items()
+    )
