+++ weblate/wladmin/tests.py
@@ -21,7 +21,7 @@ from weblate.auth.models import Group
 from weblate.trans.models import Announcement
 from weblate.trans.tests.test_views import ViewTestCase
 from weblate.trans.tests.utils import get_test_file
-from weblate.utils.checks import check_data_writable
+from weblate.utils.apps import check_data_writable
 from weblate.utils.unittest import tempdir_setting
 from weblate.wladmin.models import BackupService, ConfigurationError, SupportStatus
